+++ src/view/WindowLinha.cpp
@@ -27,10 +27,10 @@ void WindowLinha::initialize()
 WindowLinha::WinLinha WindowLinha::add_linha()
 {
 	const char* nome = gtk_entry_get_text(_entry_nome);
-	gdouble inicial_x = gtk_spin_button_get_value(_spinbutton_inicial_x);
-	gdouble inicial_y = gtk_spin_button_get_value(_spinbutton_inicial_y);
-	gdouble final_x = gtk_spin_button_get_value(_spinbutton_final_x);
-	gdouble final_y = gtk_spin_button_get_value(_spinbutton_final_y);
+	gfloat inicial_x = static_cast<float>(gtk_spin_button_get_value(_spinbutton_inicial_x));
+	gfloat inicial_y = static_cast<float>(gtk_spin_button_get_value(_spinbutton_inicial_y));
+	gfloat final_x = static_cast<float>(gtk_spin_button_get_value(_spinbutton_final_x));
+	gfloat final_y = static_cast<float>(gtk_spin_button_get_value(_spinbutton_final_y));
 	gtk_widget_hide(_window);
 	WinLinha linha = WinLinha(nome, Vector(inicial_x, inicial_y), Vector(final_x, final_y));
 	gtk_entry_set_text(_entry_nome,"" );
